requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.3.4
    # via django
billiard==3.6.4.0
    # via celery
celery[redis]==5.0.5
    # via -r requirements/requirements.in
celery-progress==0.1.0
    # via -r requirements/requirements.in
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
coreapi==2.3.3
    # via -r requirements/requirements.in
coreschema==0.0.4
    # via coreapi
cryptography==3.4.8
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-stripe==2.4.4
    # via -r requirements/requirements.in
django==3.2.8
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   djangorestframework
    #   jsonfield
django-allauth==0.44.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.2
    # via -r requirements/requirements.in
djangorestframework==3.12.4
    # via -r requirements/requirements.in
idna==2.10
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
jsonfield==3.1.0
    # via dj-stripe
kombu==5.0.2
    # via celery
markupsafe==1.1.1
    # via jinja2
oauthlib==3.1.0
    # via requests-oauthlib
prompt-toolkit==3.0.18
    # via click-repl
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.0.1
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via
    #   celery
    #   django
redis==3.5.3
    # via celery
requests==2.25.1
    # via
    #   coreapi
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
requests-oauthlib==1.3.0
    # via django-allauth
six==1.15.0
    # via click-repl
sqlparse==0.4.1
    # via django
stripe==2.56.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.4
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
dwollav2==2.2.0
plaid-python==8.6.0
python-dotenv==0.19.2
